WIP: restructure and test sorting algorithms.
